move replacer class to hpp to be used by tests create templates for c++ tests update R tests
