+++ src/cpp/session/modules/SessionFind.cpp
@@ -111,123 +111,6 @@ private:
    int nextUpdate_;
 };
 
-// helper class used to process file replacements
-class Replacer : public boost::noncopyable
-{
-public:
-   explicit Replacer(bool ignoreCase) :
-      ignoreCase_(ignoreCase)
-   {
-   }
-
-   void replaceLiteralWithLiteral(int matchOn, int matchOff, std::string* line,
-                                  const std::string* replaceLiteral,
-                                  int* pReplaceMatchOff)
-   {
-      *line = line->replace(matchOn, (matchOff - matchOn), *replaceLiteral);
-      *pReplaceMatchOff = matchOn + replaceLiteral->size();
-   }
-
-   Error replaceLiteralWithRegex(int matchOn, int matchOff, std::string* line,
-                                const std::string* replaceRegex,
-                                int* pReplaceMatchOff)
-   {
-      std::string find(line->substr(matchOn, matchOff));
-      return (replaceRegexWithRegex(matchOn, matchOff, line, &find,
-                                    replaceRegex, pReplaceMatchOff));
-   }
-
-   Error replaceRegexWithLiteral(int matchOn, int matchOff, std::string* line,
-                                 const std::string* findRegex, const std::string* replaceLiteral,
-                                 int* pReplaceMatchOff)
-   {
-      return (replaceRegexWithRegex(matchOn, matchOff, line, findRegex, replaceLiteral,
-                                    pReplaceMatchOff));
-   }
-
-  Error replaceRegexWithRegex(int matchOn, int matchOff, std::string* line,
-                              const std::string* findRegex, const std::string* replaceRegex,
-                              int* pReplaceMatchOff)
-  {
-     Error error;
-     if (ignoreCase_)
-        error = replaceRegexWithCaseInsensitiveRegex(matchOn, matchOff, line, findRegex, replaceRegex,
-                                                     pReplaceMatchOff);
-     else
-        error = replaceRegexWithCaseSensitiveRegex(matchOn, matchOff, line, findRegex, replaceRegex,
-                                                   pReplaceMatchOff);
-     return error;
-  }
-
-private:
-   Error replaceRegexWithRegex(const boost::regex* searchRegex, const boost::regex* replaceRegex,
-                               int matchOn, int matchOff, std::string* line,
-                               int* pReplaceMatchOff)
-   {
-      std::string temp;
-      try
-      {
-         temp = boost::regex_replace(*line, *searchRegex, *replaceRegex);
-      }
-      catch (const std::runtime_error& e)
-      {
-         return Error(-1, e.what(), ERROR_LOCATION);
-      }
-
-      std::string endOfString = line->substr(matchOff).c_str();
-      size_t replaceMatchOff;
-      if (endOfString.empty())
-         replaceMatchOff = temp.length();
-      else
-         replaceMatchOff = temp.find(endOfString);
-      *line = temp;
-      std::string replaceString = temp.substr(matchOn, (replaceMatchOff - matchOn));
-      *pReplaceMatchOff = matchOn  + replaceString.size();
-      return Success();
-   }
-
-   Error replaceRegexWithCaseInsensitiveRegex(int matchOn, int matchOff, std::string* line,
-                                              const std::string* findRegex, const std::string* replaceRegex,
-                                              int* pReplaceMatchOff)
-   {
-      try
-      {
-         boost::regex find(findRegex->c_str(), boost::regex::grep | boost::regex::icase);
-         boost::regex replace(replaceRegex->c_str(), boost::regex::grep | boost::regex::icase);
-
-         Error error = replaceRegexWithRegex(&find, &replace, matchOn, matchOff, line,
-                                             pReplaceMatchOff);
-         return error;
-      }
-      catch (const std::runtime_error& e)
-      {
-         return Error(-1, e.what(), ERROR_LOCATION);
-      }
-   }
-
-   Error replaceRegexWithCaseSensitiveRegex(int matchOn, int matchOff, std::string* line,
-                                            const std::string* findRegex, const std::string* replaceRegex,
-                                            int* pReplaceMatchOff)
-   {
-      try
-      {
-         boost::regex find(findRegex->c_str(), boost::regex::grep);
-         boost::regex replace(replaceRegex->c_str(), boost::regex::grep);
-
-         Error error = replaceRegexWithRegex(&find, &replace, matchOn, matchOff, line,
-                                             pReplaceMatchOff);
-         return error;
-      }
-      catch (const std::runtime_error& e)
-      {
-         return Error(-1, e.what(), ERROR_LOCATION);
-      }
-   }
-                                           
-private:
-   bool ignoreCase_;
-};
-
 // Reflects the current set of Find results that are being
 // displayed, in case they need to be re-fetched (i.e. browser
 // refresh)
